// vite-frontend/src/company.d.ts
export interface CompanyRealtimePrice {
    bidSize: number;
    askPrice: number;
    volume: number;
    askSize: number;
    bidPrice: number;
    lastSalePrice: number;
    lastSaleSize: number;
    lastSaleTime: number;
    fmpLast: number;
    lastUpdated: number;
    symbol: stringnumber;
}